signalapp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "signalapp"
version = "0.1.0"
description = "Post-call behavioral intelligence for sales teams"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Signal Team", email = "team@signal.ai"}
]

dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.35",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.1.0",

    # LangGraph Pipeline
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=2.0.0",

    # LLM Providers
    "google-genai>=0.8.0",
    "anthropic>=0.38.0",
    "openai>=1.57.0",
    "instructor>=1.6.0",

    # ASR Providers
    "aiohttp>=3.10.0",
    "assemblyai>=0.9.0",

    # ARQ Job Queue
    "arq>=0.26.0",
    "redis>=5.2.0",

    # Observability
    "langfuse>=2.0.0",

    # Storage
    "boto3>=1.35.0",

    # Configuration
    "python-dotenv>=1.0.0",

    # HTTP
    "httpx>=0.27.0",

    # Utilities
    "python-multipart>=0.0.12",
    "aiofiles>=0.8.0",

    # Streamlit (for testing harness)
    "streamlit>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["signalapp*"]

[tool.setuptools.package-dir]
"" = "."